Created a dict thats dict[team][player]
The dict holds two lists.  The first is [game time, points diff] and holds data
on every possession that player is on the court.  The second is [points diff,
points scored] and holds data each time a player gets buckets.  The plan is to
use this to create a better weighted histogram for each player.