started work on adjusting values
i now get all of the times a player is on the court, sort them by point diff,
and devide them by the total number of possesions.  this gives me an array
that i can use to adjust data for point diffs.

with that, i'm now plotting stacked bar charts of histogram data i'm creating
and then adjusting.